% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_by_fix_date}
\alias{filter_by_fix_date}
\title{Filter forecasts by fixing an issuance date (all horizons from that issue)}
\usage{
filter_by_fix_date(forecasts, issuance_date, issuance_col = "issuance_date")
}
\arguments{
\item{forecasts}{Tibble with at least columns: \code{issuance_date} and \code{horizon}}

\item{issuance_date}{Date or character parsable by lubridate}

\item{issuance_col}{Name of issuance date column (default \code{issuance_date})}
}
\value{
Filtered tibble
}
\description{
Filter forecasts by fixing an issuance date (all horizons from that issue)
}
