% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_by_fix_horizon}
\alias{filter_by_fix_horizon}
\title{Filter forecasts by fixing a horizon across multiple issuance dates}
\usage{
filter_by_fix_horizon(forecasts, horizon, horizon_col = "horizon")
}
\arguments{
\item{forecasts}{Tibble with at least \code{horizon} and \code{issuance_date}}

\item{horizon}{Integer horizon value to filter}

\item{horizon_col}{Name of horizon column (default \code{horizon})}
}
\value{
Filtered tibble
}
\description{
Filter forecasts by fixing a horizon across multiple issuance dates
}
