% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_crosswalks.R
\name{load_crosswalk}
\alias{load_crosswalk}
\title{Load a crosswalk CSV (e.g., FIPS to names, HSA mappings)}
\usage{
load_crosswalk(path)
}
\arguments{
\item{path}{Path to a CSV file}
}
\value{
tibble crosswalk
}
\description{
Load a crosswalk CSV (e.g., FIPS to names, HSA mappings)
}
