% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dashboard.R
\name{plot_dashboard_observed_vs_forecast}
\alias{plot_dashboard_observed_vs_forecast}
\title{Plot observed vs forecasts with metrocast theme and labels}
\usage{
plot_dashboard_observed_vs_forecast(observed, forecasts, facet_by = NULL)
}
\arguments{
\item{observed}{Observed tibble}

\item{forecasts}{Forecast tibble}

\item{facet_by}{Optional column to facet by}
}
\value{
ggplot object
}
\description{
Plot observed vs forecasts with metrocast theme and labels
}
