% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dashboard.R
\name{read_dashboard_series_csv}
\alias{read_dashboard_series_csv}
\title{Read dashboard series CSV}
\usage{
read_dashboard_series_csv(path)
}
\arguments{
\item{path}{Input CSV path}
}
\value{
tibble with dashboard series columns
}
\description{
Read dashboard series CSV
}
