% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{read_hub_forecast_dir}
\alias{read_hub_forecast_dir}
\title{Read and standardize all Forecast Hub CSVs in a directory}
\usage{
read_hub_forecast_dir(directory, recursive = TRUE)
}
\arguments{
\item{directory}{Directory containing CSVs}

\item{recursive}{Recurse into subdirectories}
}
\value{
tibble with bound and standardized rows
}
\description{
Read and standardize all Forecast Hub CSVs in a directory
}
