% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_quantile.R
\name{theme_metrocast}
\alias{theme_metrocast}
\title{Metrocast ggplot theme}
\usage{
theme_metrocast()
}
\value{
a ggplot2 theme
}
\description{
A clean minimal theme with improved grid and text for dashboards.
}
\examples{
# p + theme_metrocast()
}
